requirements.txt
alpaca-trade-api==3.0.0
pandas==2.0.3
numpy==1.24.3
yfinance==0.2.18
ta-lib==0.4.24
python-dotenv==1.0.0
schedule==1.2.0
requests==2.31.0
matplotlib==3.7.1
plotly==5.15.0
scikit-learn==1.3.0
datetime
logging
